Moved AbstractGlobalActionContext, need to be accessed from vsnetbeans.
