+++ java/java.lsp.server/nbproject/project.properties
@@ -17,7 +17,7 @@
 
 javac.source=1.8
 javac.compilerargs=-Xlint -Xlint:-serial
-spec.version.base=1.12.0
+spec.version.base=1.13.0
 requires.nb.javac=true
 lsp.build.dir=vscode/nbcode
 test.unit.run.cp.extra=
